pyproject.toml
[project]
name = "iivf-chatbot-backend"
version = "0.1.0"
description = "AI-powered backend for IVF chatbot"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # --- Core Framework ---
    "fastapi==0.118.3",
    "uvicorn[standard]==0.37.0",
    "starlette==0.48.0",

    # --- Database & ORM ---
    "beanie==2.0.0",
    "motor==3.7.1",
    "pymongo==4.15.3",

    # --- Config & Models ---
    "pydantic[email]==2.12.0",
    "pydantic-core==2.41.1",
    "pydantic-settings==2.11.0",

    # --- Authentication & Security ---
    "bcrypt==5.0.0",
    "passlib[argon2,bcrypt]==1.7.4",
    "pyjwt==2.10.1",

    # --- AWS & Cloud ---
    "boto3==1.40.49",
    "botocore==1.40.49",
    "s3transfer==0.14.0",

    # --- Logging & Utils ---
    "loguru==0.7.3",
    "rich==14.2.0",
    "colorama==0.4.6",

    # --- Caching & Messaging ---
    "redis==6.4.0",

    # --- Geolocation ---
    "geopy==2.4.1",
    "haversine==2.9.0",

    # --- LLM & AI Integrations ---
    "openai==1.109.1",
    "google-generativeai==0.8.5",
    "chromadb==1.1.0",
    "tiktoken==0.12.0",
    "langchain==0.3.27",
    "langchain-community==0.3.29",
    "langchain-core==0.3.76",
    "langchain-text-splitters==0.3.11",
    "langchain-openai==0.3.33",
    "langsmith==0.4.30",

    # --- PDF / Doc Processing ---
    "pypdf==3.14.0",
    "lxml==6.0.2",
    "python-docx>=1.1.0",

    # --- Google APIs ---
    "google-api-core==2.25.2",
    "google-api-python-client==2.184.0",
    "google-auth==2.41.1",
    "google-auth-httplib2==0.2.0",
    "googleapis-common-protos==1.70.0",

    # --- Misc Utilities ---
    "requests==2.32.5",
    "python-dotenv==1.1.1",
    "aiofiles>=23.2.1",
    "aiohttp==3.13.0",
    "jq==1.10.0",
    "python-multipart==0.0.20",
    "google-genai==1.43.0",
    "fastapi-mail==1.5.0",
    "apscheduler==3.11.0",
]